# docs/swagger.yaml
basePath: /
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Appointment:
    properties:
      createdAt:
        type: string
      customer:
        $ref: '#/definitions/models.User'
      customer_id:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_recurring:
        type: boolean
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      recur_pattern:
        allOf:
        - $ref: '#/definitions/models.Recurrence'
        description: E.g., "weekly", "monthly", "daily"
      recur_pattern_id:
        type: integer
      service:
        $ref: '#/definitions/models.Service'
      service_id:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/models.AppointmentStatus'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  models.AppointmentStatus:
    enum:
    - pending
    - confirmed
    - canceled
    - completed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusCanceled
    - StatusCompleted
...
  models.DayOfWeek:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    type: integer
    x-enum-varnames:
    - Sunday
    - Monday
    - Tuesday
    - Wednesday
    - Thursday
    - Friday
    - Saturday
  models.Permission:
    properties:
      action:
        description: e.g., "create", "read", "update", "delete"
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        description: e.g., "appointments", "users", etc.
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
...
  models.Recurrence:
    properties:
      appointment_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_after:
        description: Number of occurrences
        type: integer
      frequency:
        description: '"daily", "weekly", "monthly"'
        type: string
      id:
        type: integer
      next_run:
        type: string
      updatedAt:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
  models.Service:
    properties:
      buffer_time:
        allOf:
        - $ref: '#/definitions/time.Duration'
        description: Time between appointments
      cost:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      discount:
        description: Discount percentage
        type: number
      discounted_price:
        type: number
      duration:
        $ref: '#/definitions/time.Duration'
      id:
        type: integer
      name:
        type: string
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      appointments:
        items:
          $ref: '#/definitions/models.Appointment'
        type: array
      created_at:
        type: string
      customer_appointments:
        items:
          $ref: '#/definitions/models.Appointment'
        type: array
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      name:
        type: string
      otp:
        type: string
      otp_expires_at:
        type: string
      password:
        type: string
      provided_services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      working_hours:
        items:
          $ref: '#/definitions/models.WorkingHours'
        type: array
    type: object
...
  models.WorkingHours:
    properties:
      break_end:
        description: Optional break end time
        type: string
      break_start:
        description: Optional break start time
        type: string
      createdAt:
        type: string
      day_of_week:
        $ref: '#/definitions/models.DayOfWeek'
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_time:
        description: Format "HH:MM" in 24h
        type: string
      id:
        type: integer
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      start_time:
        description: Format "HH:MM" in 24h
        type: string
      updatedAt:
        type: string
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
...
host: localhost:8000
info:
  contact: {}
  description: This is the API documentation for the Appointment App.
  title: Appointment App API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access + refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: data
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login a user
      tags:
      - Auth
  /auth/logout:
    post:
      description: Logs out user (stateless in JWT; no actual invalidation)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/me:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Auth
  /auth/otp/verify/:
    post:
      consumes:
      - application/json
      description: Verifies OTP sent to the email
      parameters:
      - description: Optional action e.g. reset
        in: query
        name: action
        type: string
      - description: OTP verification data
        in: body
        name: data
        required: true
        schema:
          properties:
            email:
              type: string
            otp:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generates a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: data
        required: true
        schema:
          properties:
            refreshToken:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a user with name, email, and password. Assigns a default
        role if not provided.
      parameters:
      - description: User info
        in: body
        name: data
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /auth/reset-password/{token}:
    post:
      consumes:
      - application/json
      description: Resets the password using a token from OTP verification
      parameters:
      - description: Reset token
        in: path
        name: token
        required: true
        type: string
      - description: Password reset data
        in: body
        name: data
        required: true
        schema:
          properties:
            email:
              type: string
            new_password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - Auth
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: Sends OTP to the provided email for verification
      parameters:
      - description: Email for OTP
        in: body
        name: data
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send OTP
      tags:
      - Auth
  /auth/user/{id}:
    get:
      description: Returns user information for given ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"